feat(master): Emit signals on detained inode changes
The section FREE in the metadata is related to the IdPoolDetainer
instance in gMetadata. This commit emits a signal whenever an Id is
added or removed from detention (FREE node). The default metadata
backend (based on a file) just ignores those signals, but the FDB
backend immediately update the changes in the database.

Signed-off-by: guillex <guillex@leil.io>