refactor(master): Add the FilesystemOperationContext class
Some backends needs a transaction that starts at the beginning of the
FS operation and then propagates the same transaction to the inner
function calls, in order to use the same transaction for the complete
operation.

This commit introduces the class FilesystemOperationContext to carry on
with the transaction. The next steps will modify the needed function
signatures for the operations that will use it.

The default backend (in-memory + file based) will just ignore the
context.

Signed-off-by: guillex <guillex@leil.io>