feat: Add support for unicode when no console
Previosusly implemented versions for print_unicode_console and
print_unicode_console_error functions did not handle the case when
there was no console attached on Windows (for example, when output is
redirected to a file or when running as a background process). This
commit introduces new Windows API based implementation for managing
those scenarios.

Signed-off-by: rolysr <rolysr@leil.io>